Migrate subcommand diff to the new agent mode
